backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.19
aiofiles==24.1.0
google-generativeai==0.8.4
requests==2.32.3
chromadb==0.5.23
httpx==0.28.1
sse-starlette==2.2.1
